Refactor: Moved command handlers to separate files in commands folder
