--- .server_pid
@@ -1,1 +1,0 @@
-9341
