+++ Makefile
@@ -13,6 +13,18 @@ SOURCES = \
 	./src/Channel.cpp\
 	./src/replies.cpp\
 	./src/utils.cpp\
+	./commands/invite.cpp\
+	./commands/join.cpp\
+	./commands/kick.cpp\
+	./commands/mode.cpp\
+	./commands/part.cpp\
+	./commands/ping.cpp\
+	./commands/privmsg.cpp\
+	./commands/quit.cpp\
+	./commands/registration.cpp\
+	./commands/topic.cpp\
+	
+	
 
 
 OBJECTS = $(SOURCES:.cpp=.o)
