+++ commands/invite.cpp
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   invite.cpp                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 11:20:10 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:57:04 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+void handleInvite(Server &server, Client &client, const std::vector<std::string> &params) {
+	(void)server;
+	(void)client;
+	(void)params;
+}
