+++ commands/join.cpp
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   join.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 10:36:11 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:58:22 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+
+void handleJoin(Server &server, Client &client, const std::vector<std::string> &params)
+{
+	
+	(void)server;
+	(void)client;
+	(void)params;
+
+}
