+++ commands/kick.cpp
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   kick.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 11:19:18 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:58:55 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+
+void handleKick(Server &server, Client &client, const std::vector<std::string> &params) {
+	(void)server;
+	(void)client;
+	(void)params;
+}
