+++ commands/mode.cpp
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   mode.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 11:20:59 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:59:17 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+
+void handleMode(Server &server, Client &client, const std::vector<std::string> &params) {
+	(void)server;
+	(void)client;
+	(void)params;
+}
