+++ commands/part.cpp
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   part.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 11:22:23 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:58:33 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+
+void handlePart(Server &server, Client &client, const std::vector<std::string> &params) {
+	(void)server;
+	(void)client;
+	(void)params;
+}
