+++ commands/ping.cpp
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ping.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 10:36:19 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:58:14 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+
+void handlePing(Server &server, Client &client, const std::vector<std::string> &params)
+{
+	(void)server;
+	(void)client;
+	(void)params;
+}
