+++ commands/quit.cpp
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   quit.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 11:17:52 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:58:26 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+
+void handleQuit(Server &server, Client &client, const std::vector<std::string> &params)
+{
+	(void)server;
+	(void)client;
+	(void)params;
+}
