+++ commands/topic.cpp
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   topic.cpp                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/04 11:22:07 by akurmyza          #+#    #+#             */
+/*   Updated: 2025/07/04 11:58:48 by akurmyza         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/commands.hpp"
+void handleTopic(Server &server, Client &client, const std::vector<std::string> &params) {
+	(void)server;
+	(void)client;
+	(void)params;
+}
