+++ include/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/15 17:31:22 by akurmyza          #+#    #+#             */
-/*   Updated: 2025/07/03 19:13:51 by akurmyza         ###   ########.fr       */
+/*   Updated: 2025/07/04 12:00:29 by akurmyza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,11 +35,12 @@
 #include <fcntl.h>      // fcntl()
 
 //======================== PROJECT HEADERS ===================================//
-#include "../include/colors.hpp"
+#include "../include/commands.hpp"
 #include "../include/utils.hpp"
 #include "../include/replies.hpp"
 #include "../include/Client.hpp"
 #include "../include/Channel.hpp"
+#include "../include/colors.hpp"
 
 //======================== CLASS DEFINITION ==================================//
 class Server
@@ -60,66 +61,45 @@ private:
     std::map<int, Client *> _clients;
     std::map<std::string, Channel *> _channels;
 
-    //======================== PRIVATE: CONSTRUCTORS  (only 1 server) ==============//
+    //======================== PRIVATE: CONSTRUCTORS (Singleton-style prevention) ==//
     Server();
     Server(const Server &o);
     Server &operator=(const Server &o);
 
-    //======================== PRIVATE: HELPER FUNCTIONS =======================//
-
-    // ─────────────────── Connection & Client Management ───────────────────
-    bool isNicknameInUse(const std::string &nickname);
+    //======================== PRIVATE: CONNECTION & CLIENT MANAGEMENT =============//
+   
     Client *getClientByNickname(const std::string &nickname);
-    static bool isValidNickname(const std::string &nickname);
-    static bool isValidUsername(const std::string &username);
-    
     void acceptNewClient();
     void removeClient(int fd, size_t pollFdIndex);
     void handleClientInput(size_t pollFdIndex);
-    void sendToClient(int fd, const std::string &message);
 
-    // ─────────────────── Channel Management ───────────────────────
+    //======================== PRIVATE: CHANNEL MANAGEMENT =========================//
     Channel *getChannelByName(const std::string &channelName);
     bool isChannelName(const std::string &channelName);
 
-    // ─────────────────── Command Dispatch & Parsing ───────────────────
+    //======================== PRIVATE: COMMAND DISPATCH & PARSING =================//
     void handleCommand(Client *client, const std::string &line);
 
-    // ─────────────────── Mandatory IRC Command Handlers
-    void handlePass(Client *client, const std::vector<std::string> &params);
-    void handleNick(Client *client, const std::vector<std::string> &params);
-    void handleUser(Client *client, const std::vector<std::string> &params);
-    void handleJoin(Client *client, const std::vector<std::string> &params);
-    void handlePrivateMessage(Client *client, const std::vector<std::string> &params);
-
-    // ─────────────────── Internal Utilities ───────────────────
+    //======================== PRIVATE: INTERNAL UTILITIES =========================//
     void checkResult(int result, const std::string &errMsg);
     void logInfo(const std::string &msg);
     void logErrAndThrow(const std::string &msg);
 
 
-     
-
-    /* Optional for full channel logic or bonus commands
-
-    void broadcastToChannel(const std::string &channel, const std::string &msg, int except_fd = -1);
-    void handleKick(Client *client, const std::vector<std::string> &params);
-    void handleInvite(Client *client, const std::vector<std::string> &params);
-    void handleTopic(Client *client, const std::vector<std::string> &params);
-    void handleMode(Client *client, const std::vector<std::string> &params);
-    */
-
 public:
-    //======================== PUBLIC: CONSTRUCTORS & DESTRUCTORS ==============//
+    //======================== PUBLIC: CONSTRUCTORS & DESTRUCTOR ===================//
     Server(int port, const std::string &password);
     ~Server();
 
-    //======================== PUBLIC: MAIN SERVER METHODS =====================//
+    //======================== PUBLIC: MAIN SERVER METHODS =========================//
     int run();
 
-    //======================== PUBLIC: GETTERS =================================//
+    //======================== PUBLIC: CLIENT UTILITIES ============================//
+    void sendToClient(int fd, const std::string &message);
+    void sendPong(Client &client, const std::string &message);
+    bool isNicknameInUse(const std::string &nickname);
+
+    //======================== PUBLIC: GETTERS =====================================//
     int getPort() const;
     const std::string &getServerName() const;
-   
 };
-
