+++ include/utils.hpp
@@ -6,7 +6,7 @@
 /*   By: akurmyza <akurmyza@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/03 18:20:44 by akurmyza          #+#    #+#             */
-/*   Updated: 2025/07/03 18:29:31 by akurmyza         ###   ########.fr       */
+/*   Updated: 2025/07/04 10:59:10 by akurmyza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,12 @@
 #include <iostream>
 #include <sstream> // std::stringstream for intToString
 
-//======================== HELPER FUNCTIONS ========================//
+//======================== PUBLIC: NICKNAME & USERNAME VALIDATION ==================//
+bool isValidNickname(const std::string &nickname);
+bool isValidUsername(const std::string &username);
 
-bool isValidPassword(const std::string &password);  
-   
+//======================== PUBLIC: PASSWORD VALIDATION =============================//
+bool isValidPassword(const std::string &password);
+
+//======================== PUBLIC: CONVERSION UTILITIES =============================//
 std::string intToString(int n);
